pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "crawler"
version = "1.0.0"
description = "Playwright crawler sharing app/ configs and SQLite"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
crawler = "crawler.__main__:main"

[tool.setuptools]
package-dir = {"crawler" = "crawler"}
packages = ["crawler"]
